# docs/swagger/openapi-converted.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API for managing community watch shifts, bookings, and reports",
        "title": "Community Watch Shift Scheduler API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/admin/bookings/assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an admin to book a specific user for a given schedule ID and start time.",
                "tags": [
                    "admin-bookings"
                ],
                "summary": "Assign a user to a specific shift slot (Admin)",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.AssignUserToShiftRequest"
                            }
                        }
                    },
                    "description": "Assignment details (schedule_id, start_time, user_id)",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully for the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.BookingResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request format or data (e.g., missing fields, invalid user/schedule)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - admin authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user does not have admin privileges",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User or Schedule not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Slot already booked or other conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/users/{userId}/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an admin to view all bookings (past and future) for a specific user.",
                "tags": [
                    "admin-bookings"
                ],
                "summary": "Get all bookings for a specific user (Admin)",
                "parameters": [
                    {
                        "example": 42,
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings for the user with schedule names",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/api.BookingWithScheduleResponse"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - admin authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user does not have admin privileges",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/bookings": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Books a shift slot for a user",
                "tags": [
                    "bookings"
                ],
                "summary": "Create a new booking",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.CreateBookingRequest"
                            }
                        }
                    },
                    "description": "Booking details",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.BookingResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request format or data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Schedule not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Slot already booked (conflict)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/bookings/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel a user's booking if it's not too close to the shift start time",
                "tags": [
                    "bookings"
                ],
                "summary": "Cancel a booking",
                "parameters": [
                    {
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Booking cancelled successfully"
                    },
                    "400": {
                        "description": "Invalid booking ID or booking cannot be cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to cancel this booking",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://localhost:5888"
        }
    ],
    "components": {
        "schemas": {
...
            "api.AssignUserToShiftRequest": {
                "type": "object",
                "properties": {
                    "schedule_id": {
                        "type": "integer"
                    },
                    "start_time": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
            "api.BookingResponse": {
                "type": "object",
                "properties": {
                    "booking_id": {
                        "type": "integer"
                    },
                    "buddy_name": {
                        "type": "string"
                    },
                    "buddy_user_id": {
                        "type": "integer"
                    },
                    "checked_in_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "schedule_id": {
                        "type": "integer"
                    },
                    "shift_end": {
                        "type": "string"
                    },
                    "shift_start": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
            "api.BookingWithScheduleResponse": {
                "type": "object",
                "properties": {
                    "booking_id": {
                        "type": "integer"
                    },
                    "buddy_name": {
                        "type": "string"
                    },
                    "buddy_user_id": {
                        "type": "integer"
                    },
                    "checked_in_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "schedule_id": {
                        "type": "integer"
                    },
                    "schedule_name": {
                        "type": "string"
                    },
                    "shift_end": {
                        "type": "string"
                    },
                    "shift_start": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
            "api.CreateBookingRequest": {
                "type": "object",
                "properties": {
                    "buddy_name": {
                        "type": "string"
                    },
                    "buddy_phone": {
                        "type": "string"
                    },
                    "schedule_id": {
                        "type": "integer"
                    },
                    "start_time": {
                        "type": "string"
                    }
                }
            },
...
            "api.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}